Refactor physics module

**Summary**:  
*  Modifies the Dynamical objects realted classes to generalize for any problem state. This allows for rigid bodies, deformable bodies using FEM and pbd under one dynamicalObject base class.
* Adds common interface for pbd solver and FE solvers to allow simpler code in SceneManager
* Refactors all the PBD classes for bugs, style, warnings and efficiency.
* Misc. modifications including rearranging folders, example codes, renaming classes and clearing some warnings.

See merge request !103

+++ Base/Geometry/Map/imstkOneToOneMap.h
@@ -64,7 +64,7 @@ public:
     ///
     /// \brief Sets the one-to-one correspondence directly
     ///
-    void setMap(const std::map<int, int>& sourceMap);
+    void setMap(const std::map<size_t, size_t>& sourceMap);
 
     ///
     /// \brief Apply (if active) the tetra-triangle mesh map
@@ -86,12 +86,13 @@ public:
     ///
     void setSlave(std::shared_ptr<Geometry> slave) override;
 
-    unsigned int getMapIdx(const unsigned int& idx)
-    {
-        return m_oneToOneMap[idx];
-    }
+    ///
+    /// \brief
+    ///
+    size_t getMapIdx(const size_t& idx) { return m_oneToOneMap[idx]; }
+
 protected:
-    std::map<int, int> m_oneToOneMap; ///> One to one mapping data
+    std::map<size_t, size_t> m_oneToOneMap; ///> One to one mapping data
 };
 
 } // imstk
